[feature] TestCase 3개 추가 구현

+++ Project22/similarity.cpp
@@ -12,15 +12,16 @@ public:
 		if (len1 == len2) return 60;
 		if (len1 == len2 * 2 || len1 * 2 == len2) return 0;
 
-		int gap, min;
+		 double gap, min;
 		if (len1 > len2) {
 			gap = len1 - len2;
 			min = len2;
+			
 		}
 		else {
 			gap = len2 - len1;
 			min = len1;
 		}
-		return (1 - gap / min) * 60;
+		return (1- gap / min) * 60;
 	}
 };
